ai-data-analyst/requirements.txt
langchain
langchain-community
langchain-google-genai
pydantic
python-dotenv
matplotlib
pandas
sqlalchemy